adds and corrects tests
